+++ src/python/zensols/db/__init__.py
@@ -1,2 +1,3 @@
 from .parse import *
 from .bean import *
+from .stash import *